Fix QScrollBar Stylesheet
By convention in the recent releases of Flame, the scroll bar background
should be a darker grey and the handle would be a lighter grey.
Something changed in PySide6 and it was no longer rendering correctly.

Add style rules for add-page and sub-page, which are meant to be unique
rules for parts of the scroll bar that have not been visited yet.  It
was adding a light grey checkboard pattern over the scroll bar which
made it look too similar to the handle.

See article below
https://forum.qt.io/topic/134913/unwanted-checkerboard-pattern-on-background-of-qscrollbar/7